+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 
 in progress
 ===========
+- :ref:`firmware-builder`: Don't pull **all** build parameters into the artefact filename
+  as this might raise ``OSError: [Errno 63] File name too long``.
 
 
 .. _kotori-0.16.0: